performance-meter/edge-agent/requirements.txt
paho-mqtt==1.5.1

requests~=2.22.0
stomper~=0.4.3

websocket-client~=0.58.0
